Archive_tables/requirements.txt
# requirements.txt
python-dotenv==1.0.0
azure-storage-blob==12.15.0
pandas>=1.3.0
pyodbc>=4.0.30
tqdm>=4.62.0